Exercise 4.21 - bloglist expansion Step9

+++ controllers/blogs.js
@@ -1,7 +1,6 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
-const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', ['name', 'username', 'id'])
@@ -10,14 +9,9 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
+  const authUserId = request.authUserId
 
-  const decodedToken = jwt.verify(request.token, process.env.SECRET)
-
-  if (!decodedToken.id) {
-    return response.status(401).json({ error: 'token invalid' })
-  }
-
-  const user = await User.findById(decodedToken.id)
+  const user = await User.findById(authUserId)
 
   const blog = new Blog({
     title: body.title,
@@ -36,12 +30,20 @@ blogsRouter.post('/', async (request, response) => {
   response.json(savedBlog)
 })
 
-blogsRouter.delete('/:id', (request, response, next) => {
-  Blog.findByIdAndRemove(request.params.id)
-    .then((result) => {
-      response.status(204).json(result)
-    })
-    .catch((error) => next(error))
+blogsRouter.delete('/:id', async (request, response, next) => {
+  const authUserId = request.authUserId
+  const blog = await Blog.findById(request.params.id)
+  const userId = blog.user.toString()
+
+  if (authUserId === userId) {
+    Blog.findByIdAndRemove(request.params.id)
+      .then((result) => {
+        response.status(204).json(result)
+      })
+      .catch((error) => next(error))
+  } else {
+    response.status(401).json('User is not authorized to delete this blog')
+  }
 })
 
 blogsRouter.put('/:id', (request, response, next) => {
